+++ core/langs/bnf/lang/__init__.py
@@ -1,0 +1,4 @@
+"""Init module for `bnf` lang."""
+from .models import *
+from .reader import *
+