+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `base/building` which implement the builder function that can construct an object from a `base.processing.Element`
   object
 - `core/langs` package which contains all the required langs of the project.
+- `core/langs/bnf` which implement the initial version for
+  the [bnf](https://en.wikipedia.org/wiki/Backus%E2%80%93Naur_form)-like language package.
 
 ## [0.0.1](https://github.com/GabrielAmare/Language/releases/tag/v0.0.1) - 2022-06-07